move trajectory without ros outside of simpl_trajectory_publisher file
